+++ R/nas.R
@@ -2,15 +2,64 @@
 #'
 #' Remove NAs from a vector
 #'
-#' @param x A vector
-#' @return `x` without values where `is.na(x)` is `TRUE`
+#' @details
+#' `remove_na.factor` will remove `NA` values as identified by the `levels()`
+#'   or by the integer value of the level.  `factors` are recreated with all
+#'   `NA` values and, if present, the `NA` `level` removed.
+#'
+#' @param x A vector of values
+#' @returns
+#'   `x` without values where `is.na(x)` is `TRUE`
+#'   For factors, a new factor (`ordered` if `is.ordered(x)`)
 #' @export
+#' @examples
+#' remove_na(c(4, 1, 2, NA, 4, NA, 3, 2))
+#'
+#' # removes based on levels
+#' remove_na(fact(c("b", NA, "a", "c")))
+#'
+#' # removes based on values
+#' x <- as_ordered(c("b", "d", "a", "c"))
+#' x[2:3] <- NA
+#' str(remove_na(x))
 remove_na <- function(x) {
-  if (!is.vector(x)) {
-    stop("x must be a vector", call. = FALSE)
+  check_is_vector(x, "any")
+  UseMethod("remove_na", x)
+}
+
+#' @rdname remove_na
+#' @export
+remove_na.default <- function(x) {
+  x[!is.na(x)]
+}
+
+#' @rdname remove_na
+#' @export
+remove_na.list <- function(x) {
+  lapply(x, remove_na)
+}
+
+#' @rdname remove_na
+#' @export
+remove_na.factor <- function(x) {
+  lvls <- levels(x)
+  na_levels <- is.na(lvls)
+
+  out <- x[!is.na(x)]
+  out <- as.integer(out)
+
+  if (any(na_levels)) {
+    which_na_level <- which(na_levels)
+    out <- out[out != which_na_level]
+
+    if (which_na_level != length(lvls)) {
+      out <- match(out, unique(out))
+    }
   }
 
-  x[!is.na(x)]
+  levels(out) <- lvls[!na_levels]
+  class(out) <- c(if (is.ordered(x)) "ordered", "factor")
+  out
 }
 
 unique_no_na <- function(x) {
@@ -29,7 +78,7 @@ unique_no_na <- function(x) {
 #' remove_null(x)
 #' @export
 remove_null <- function(x) {
-  if (!is.vector(x, "list")) {
+  if (!inherits(x, "list")) {
     stop("x must be a list", call. = FALSE)
   }
 
@@ -78,4 +127,40 @@ is_na_cols <- function(x, names = TRUE) {
   }
 
   vap_lgl(x, function(xx) all(is.na(xx)), .nm = names)
+}
+
+#' Table NA values
+#'
+#' Tables out whether data are NAs are not
+#'
+#' @details
+#' All data are checked with `is.na()` and the resulting `TRUE` or `FALSE` is
+#'   are tabulated.
+#'
+#' @inherit base::table
+#' @param .list Logical, if `TRUE` and `...` is a `list`, will c
+#' @export
+#' @examples
+#' x <- list(
+#'   a = c(1, 2, NA, 3),
+#'   b = c("A", NA, "B", "C"),
+#'   c = as.Date(c("2020-01-02", NA, NA, "2020-03-02"))
+#' )
+#' tableNA(x) # entire list
+#' tableNA(x, .list = TRUE) # counts for each
+#' tableNA(x[1], x[2])
+#' tableNA(x[1], x[2], x[3]) # equivalent ot tableNA(x, .list = TRUE)
+
+tableNA <- function(..., .list = FALSE) {
+  ls <- if (.list) {
+    as.list(...)
+  } else {
+    list(...)
+  }
+
+  if (is.null(names(ls))) {
+    names(ls) <- as.character(sys.call())[-1]
+  }
+
+  table(lapply(ls, function(x) mark::fact(is.na(x))))
 }