+++ tests/testthat/test-sourcing.R
@@ -29,13 +29,13 @@ test_that("Rscript", {
   # skip_if_not(.Platform$OS.type == "windows")
 
   x <- test_path("scripts", "rscript-test.R")
-  rscript(x, "vanilla", stdout = FALSE, stderr = FALSE)
+  expect_error(rscript(x, "vanilla", stdout = FALSE, stderr = FALSE), NA)
   expect_false("dplyr" %in% search())
 
-  rscript(x, stdout = FALSE, stderr = FALSE)
+  expect_error(rscript(x, stdout = FALSE, stderr = FALSE), NA)
   expect_false("dplyr" %in% search())
 
-  e <- source_to_env(x)
+  e <- source_to_env(x, ops = "vanilla")
   expect_s3_class(e, c("source_env", "environment"))
   expect_identical(e$a_litte_note, "You're doing okay")
   expect_s3_class(e$out, "tbl_df")